Refactor code for more modularity; Set back to 'choose a waifu' option when switching to random or mix modes; Clarify loadOnStart option; Mix mode properly showing themed favicon;
