+++ firefoxThemes/modules/utils/random.js
