+++ firefoxThemes/options/options.html
@@ -3,7 +3,6 @@
 <head>
   <meta charset="utf-8"/>
   <title>Doki Theme Options</title>
-
   <script src="options.js" type="module" defer></script>
   <link rel="stylesheet" href="option.css"/>
 </head>
@@ -11,7 +10,7 @@
 <div class="options-container">
   <section>
     <span id="loadOnStart" role="checkbox"></span>
-    <label for="loadOnStart">Load New Tab on new window</label>
+    <label for="loadOnStart">Load Doki Theme page on startup</label>
   </section>
   <section>
     <span id="textSelection" class="checked" role="checkbox"></span>
