+++ firefoxThemes/waifus/index.html
@@ -12,7 +12,7 @@
     <!--Adds/Removes Waifu-->
     <script src="waifu.js" defer></script>
   <!--Load Themed Favicon-->
-  <script type="module" src="iconloader.js" defer></script>
+  <script type="module" src="faviconLoader.js" defer></script>
 </head>
 <body>
 </body>
