+++ firefoxThemes/waifus/tab.js
@@ -1,4 +1,4 @@
-import {svgToPng, buildSVG} from "../modules/logo.js";
+import {svgToPng, buildSVG} from "../modules/utils/themes/logo.js";
 
 let query = "";//the search query
 /*Record the keywords to search for.
@@ -50,18 +50,10 @@ function setThemedAboutIcon(currentTheme) {
   });
 }
 
-function applyTabListeners() {
-  browser.storage.local.get(["showWidget", "waifuThemes", "currentThemeId"])
-    .then((storage) => {
-      const currentTheme = storage.waifuThemes.themes[storage.currentThemeId] ||
-        storage.waifuThemes.themes["19b65ec8-133c-4655-a77b-13623d8e97d3"]; // Ryuko Dark
-      const root = document.querySelector(':root');
-      root.style.setProperty('--accent-color', currentTheme.definition.colors.accentColor);
-      root.style.setProperty('--base-background-color', currentTheme.definition.colors.baseBackground);
-
-      if (storage.showWidget === undefined || storage.showWidget) {
-        document.querySelector("body").innerHTML =
-          `
+/*Display search widget*/
+function displayWidget() {
+  document.querySelector("body").innerHTML =
+    `
     <main>
     <div class="logo-and-wordmark">
         <div class="logo"></div>
@@ -75,6 +67,19 @@ function applyTabListeners() {
     </div>
 </main>
 `;
+}
+
+function applyTabListeners() {
+  browser.storage.local.get(["showWidget", "waifuThemes", "currentThemeId"])
+    .then((storage) => {
+      const currentTheme = storage.waifuThemes.themes[storage.currentThemeId] ||
+        storage.waifuThemes.themes["19b65ec8-133c-4655-a77b-13623d8e97d3"]; // Ryuko Dark
+      const root = document.querySelector(':root');
+      root.style.setProperty('--accent-color', currentTheme.definition.colors.accentColor);
+      root.style.setProperty('--base-background-color', currentTheme.definition.colors.baseBackground);
+
+      if (storage.showWidget === undefined || storage.showWidget) {
+        displayWidget();
         // set themed icons
         setThemedSearchInputIcon(currentTheme);
         setThemedAboutIcon(currentTheme);
@@ -89,4 +94,5 @@ function applyTabListeners() {
       }
     });
 }
+
 applyTabListeners();
